Tweaks for fewer preview sizes, device rotation, and slideshow zooming.
Request preview sizes in multiples of 100.

Request preview sizes based on longest edge of window.
This allows the image to still be large enough on device rotation.

Scale images in Bigshot that are larger than window.

Set minZoom using getZoomToFitValue if the image is larger than the
window.  This allows us to downscale, but prevents the user from
zooming out (making the image smaller) more than necessary.

Moved the scaling logic to _resetZoom function.

Stop treating SVG as small images.
This fixes zoom issues for SVG images in the slideshow.