Up the memory limit for composer
The web server may be configured with a 1+GB limit, but the cli may still be limited to 512MB or less, so we need to raise the limit for composer.
Also, because there are so many packages, we need to download from source rather than from dist. It's slower, prevents caching and requires more disk space, but using dist files requires setting up oAuth
[ci skip]
(cherry picked from commit 987c8b8)