Use Math.ceil() when scaling with devicePixelRatio
When rounding off dimensions of an image after applying
devicePixelRatio Math.floor() was used. This is not a good
idea as it causes the picture to be a little blurry. This
is because they're 1 pixel smaller than the expected size.
This commit changes these occurences to Math.ceil() to avoid
this blur.

Fixes https://github.com/owncloud/gallery/issues/484
(cherry picked from commit 8c350f2)