Hide three dots menu toggle in share dropdown when not shown explicitly
The link text input is hidden when not shown explicitly through
"showLink"; the three dots menu toggle is shown inside that text input,
so it should be hidden too when the link text input is hidden.

This fixes the three dots menu toggle being visible when the share
dropdown is open and the link is not shared.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>