Fix date picker not visible in slideshow sharing dialog
The z-index for the date picker is forced in the server CSS files to
"1111" through the use of "!important", so even if a higher value is
explicitly set in the "style" attribute of the element the value set in
the server CSS files is the one used. As the z-index of the slideshow is
"100000" and the date picker is added as a sibling of it and not as a
child the date picker appears behind the slideshow. In order to be
visible the date picker needs to override the z-index set in the server
CSS files to be higher than the one used for the slideshow and for the
share dialog shown in the slideshow.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>