Put progress filling behind pause icon
The progress filling for the pause icon is a next sibling of the pause
icon in the DOM. Both had the same z-index, so the progress filling
element appeared in front of the pause icon. Due to this, if the pointer
was hovering on the pause icon the hovered element could change from the
pause icon to the progress filling when its height was increased, which
in turn caused the highlight of the pause icon to disappear even if the
pointer was not moved. Now the z-index of the progress filling was
reduced to be shown behind the pause icon instead; visually it looks the
same, but this ensures that the pause icon is always the one being
hovered.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>