Hide the share drop down when closing the slideshow
The share drop down is shared between the gallery and the slideshow.
When the share button of the gallery is clicked the share drop down is
shown only if it is not already being shown. However, if the slideshow
is closed while the drop down is being shown the gallery considers that
the drop down is being shown (even if it is not visible to the user), so
clicking on the share button does nothing in that case. Now the drop
down is explicitly hidden when the slideshow is closed to ensure that
the share button of the gallery works as expected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>