Fix loading icon position in "shareWith" field
When searching or selecting a sharee in the "shareWith" field a loading
icon is shown. However, as the icon had a relative position and the
"shareWith" field uses the full width of the share dropdown the icon was
shown in a new line below the field. Moreover, due to its "right" CSS
property the icon was also shown outside the dropdown itself. Now,
during the search for or the selection of a sharee, the loading icon
replaces the information icon shown in the "shareWith" field, just like
done in the sharing section of the server.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>