Fix adding shares to wrong HTML element
"_addShareWith" adds a share to the "#shareWithList" element, which is a
descendant of the share dropdown. However, the "Sharing" tab of the
Files app sidebar contains its own "#shareWithList" element, so when the
share dropdown is shown on top of the Files app (by clicking on an image
and then sharing it from the slideshow) and the "Sharing" tab is open in
the sidebar the share element ends added to the wrong element. To
prevent this the matched element is now limited to descendants of the
share dropdown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>