Fix shortcuts activated when typing in the share dropdown
When in slideshow mode certain keys trigger actions like toggling full
screen or zooming. However, due to event bubbling, those actions were
triggered too when typing in the input fields of the share dropdown.
Instead of stopping the propagation in each input field this is now
fixed by not triggering the actions if the element that received the key
press was an input field.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>