Fix "connectToThumbnails" API test when run in Travis CI
The server uses the Imagick PHP extension to generate PostScript
previews and, in turn, ImageMagick requires Ghostscript to handle
PostScript files. It seems that the base system used by Travis CI does
not have Ghostscript installed by default, so it must be explicitly
installed for the "connectToThumbnails" API test to work as expected.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>