Fix "SignInAsUser" acceptance test
The app menu markup has changed in the server. Before, the image had an
"alt" attribute with the name of the app, so calling "click" with the
app name was enough. However, currently menu icons no longer have that
"alt" attribute; as the "Gallery" text appears in a sibling element of
the link clicking on "Gallery" does nothing, so the link has to be
located using an XPath expression instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>