Wait for the Gallery icon to appear before clicking on it
User pages returned by the server have all the apps hidden, and once the
JavaScript is initialized they are shown or kept in the "More apps" menu
based on the available width of the page. In these tests the Gallery
icon always ends being shown, but it is necessary to wait until it is
visible before clicking on it.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>