Fix timeout in API tests
When run on PHP 7.0 and MySQL in Travis CI sometimes the API tests fail
with a timeout; it seems that the tests are slower in that environment,
so the timeout has to be increased in that case.

Increasing the timeout simply gives slower tests more time to finish the
requests, but it does not affect the other tests in any way, so the
timeout was increased for every test.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>