Fix element to append the share drop down to
The share drop down was appended to the grandparent element of the
clicked element that caused the drop down to be shown. This, in itself,
was not problematic; however, due to the CSS rules for the actions bar
when the drop down was a child of the action bar (that is, when shown in
the gallery) it was not properly shown. Now the drop down is appended to
either the "#slideshow" element (when in the slideshow, just like
before) or to the "#controls" element (when in the gallery, the parent
of the actions bar).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>