Fix position of icons in "Share with" input field
The "box-sizing" of "#app" descendants is set to "border-box" in the
server. Depending on where it is shown, the "Share with" input field can
be a descendant of "#app" or not, so its icons either use the
"border-box" or the "content-box" sizing depending on the case.
Therefore, the padding when using "border-box" sizing has to be set to
the content size ("min-width" and "min-height") plus the padding used in
"content-box" sizing.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>