Hide tooltip when the input field is cleared
The "No results found" tooltip is now removed when the input field is
cleared.

This unifies its behaviour with the one used in the server, but it does
not address its problems (for example, if a search is started and while
it is being performed the input field is cleared the tooltip would be
still shown once the search response is received, even if the input
field is now empty).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>