Fix UI while a share is being added
After calling "Gallery.Share.share" the UI was restored (the working
icon was hidden, the input field enabled...). However, "share" is
asynchronous, so the UI was restored while the share was still being
added. Now the UI is restored in the share callbacks, so it is restored
once the sharing finished, either successfully or with a failure; in
this later case now a notification is also shown.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>