Hide "No results found" tooltip on autocompletion
If the tooltip is not explicitly hidden it would still be shown once the
autocompletion dropdown is shown (behind it while it is open, but fully
visible when it is closed).

This unifies its behaviour with the one used in the server, but it does
not address its problems (for example, if a search is started and while
it is being performed the input field is cleared the tooltip would be
still shown once the search response is received, even if the input
field is now empty).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>