Fix vertical scrolling
In the layout provided by the server the scrolling container is now the
body instead of the content wrapper.

Note that the container used in "galleryfileaction.js" was not modified
on purpose, as that feature is currently broken; it restores the scroll
position when the slideshow is hidden due to the file list being
reloaded, but that does not happen in Files app, only in public pages.
Thus, in Files app it restores the position as soon as the file list is
reloaded after closing the slideshow, so it changes the scroll position,
for example, when the directory changes. However, as "#app-content" is
used as the scrolling container in that case the scroll position of the
body does not actually change.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>